Chore: refactor keyframes library
